+++ packages/auth/src/Auth.ts
@@ -166,6 +166,7 @@ export default class AuthClass {
                 },
                 onFailure: (err) => {
                     logger.debug("Error in cognito hosted auth response", err);
+                    dispatchAuthEvent('signIn_failure', err);
                 }
             };
             // if not logged in, try to parse the url.
@@ -320,7 +321,7 @@ export default class AuthClass {
                     resolve(user);
                 }
             },
-        onFailure: (err) => {
+            onFailure: (err) => {
                 logger.debug('signIn failure', err);
                 dispatchAuthEvent('signIn_failure', err);
                 reject(err);
@@ -415,9 +416,11 @@ export default class AuthClass {
                 if (err) {
                     logger.debug('get MFA Options failed', err);
                     rej(err);
+                    return;
                 }
                 logger.debug('get MFA options success', mfaOptions);
                 res(mfaOptions);
+                return;
             });
         });
     }
@@ -427,18 +430,72 @@ export default class AuthClass {
      * @param {CognitoUser} user - the current cognito user
      */
     public getPreferredMFA(user: any): Promise<string> {
+        const that = this;
         return new Promise((res, rej) => {
             user.getUserData((err, data) => {
                 if (err) {
                     logger.debug('getting preferred mfa failed', err);
-                    rej('getting preferred mfa failed: ' + err);
+                    rej(err);
+                    return;
                 }
-                const preferredMFA = data.PreferredMfaSetting || 'NOMFA';
-                res(preferredMFA);
+                
+                const mfaType = that._getMfaTypeFromUserData(data);
+                if (!mfaType) {
+                    rej('invalid MFA Type');
+                    return;
+                } else {
+                    res(mfaType);
+                    return;
+                }
             });
         });
     }
+
+    private _getMfaTypeFromUserData(data) {
+        let ret = null;
+        const preferredMFA = data.PreferredMfaSetting;
+        // if the user has used Auth.setPreferredMFA() to setup the mfa type
+        // then the "PreferredMfaSetting" would exist in the response
+        if (preferredMFA) {
+            ret = preferredMFA;
+        } else {
+            // if mfaList exists but empty, then its noMFA
+            const mfaList = data.UserMFASettingList;
+            if (!mfaList) {
+                // if SMS was enabled by using Auth.enableSMS(), 
+                // the response would contain MFAOptions
+                // as for now Cognito only supports for SMS, so we will say it is 'SMS_MFA'
+                // if it does not exist, then it should be NOMFA
+                const MFAOptions = data.MFAOptions;
+                if (MFAOptions) {
+                    ret = 'SMS_MFA';
+                } else {
+                    ret = 'NOMFA';
+                }
+            } else if (mfaList.length === 0) {
+                ret = 'NOMFA';
+            } else {
+                logger.debug('invalid case for getPreferredMFA', data);
+            }
+        }
+        return ret;
+    }
     
+    private _getUserData(user) {
+        return new Promise((res, rej) => {
+            user.getUserData((err, data) => {
+                if (err) {
+                    logger.debug('getting user data failed', err);
+                    rej(err);
+                    return;
+                } else {
+                    res(data);
+                    return;
+                }
+            });
+        });
+        
+    }
     /**
      * set preferred MFA method
      * @param {CognitoUser} user - the current Cognito user
@@ -446,27 +503,59 @@ export default class AuthClass {
      * @return - A promise resolve if success
      */
     public async setPreferredMFA(user : any, mfaMethod : string): Promise<any> {
-        let smsMfaSettings = await this.getPreferredMFA(user) === 'SMS_MFA'?
-            { PreferredMfa : false, Enabled : false } : null;
-        let totpMfaSettings = {
-            PreferredMfa : false,
-            Enabled : false
-        };
+        const userData = await this._getUserData(user);
+        let smsMfaSettings = null;
+        let totpMfaSettings = null;
 
         switch(mfaMethod) {
-            case 'TOTP':
+            case 'TOTP' || 'SOFTWARE_TOKEN_MFA':
                 totpMfaSettings = {
                     PreferredMfa : true,
                     Enabled : true
                 };
                 break;
-            case 'SMS':
+            case 'SMS' || 'SMS_MFA':
                 smsMfaSettings = {
                     PreferredMfa : true,
                     Enabled : true
                 };
                 break;
             case 'NOMFA':
+                const mfaList = userData['UserMFASettingList'];
+                const currentMFAType = await this._getMfaTypeFromUserData(userData);
+                if (currentMFAType === 'NOMFA') {
+                    return Promise.resolve('No change for mfa type');
+                } else if (currentMFAType === 'SMS_MFA') {
+                    smsMfaSettings = {
+                        PreferredMfa : false,
+                        Enabled : false
+                    };
+                } else if (currentMFAType === 'SOFTWARE_TOKEN_MFA') {
+                    totpMfaSettings = {
+                        PreferredMfa : false,
+                        Enabled : false
+                    };
+                } else {
+                    return Promise.reject('invalid MFA type');
+                }
+                // if there is a UserMFASettingList in the response
+                // we need to disable every mfa type in that list
+                if (mfaList && mfaList.length !== 0) {
+                    // to disable SMS or TOTP if exists in that list
+                    mfaList.forEach(mfaType => {
+                        if (mfaType === 'SMS_MFA') {
+                            smsMfaSettings = {
+                                PreferredMfa : false,
+                                Enabled : false
+                            };
+                        } else if (mfaType === 'SOFTWARE_TOKEN_MFA') {
+                            totpMfaSettings = {
+                                PreferredMfa : false,
+                                Enabled : false
+                            };
+                        }
+                    });
+                }
                 break;
             default:
                 logger.debug('no validmfa method provided');
@@ -474,48 +563,23 @@ export default class AuthClass {
         }
 
         const that = this;
-        const TOTP_NOT_VERIFED = 'User has not verified software token mfa';
-        const TOTP_NOT_SETUP = 'User has not set up software token mfa';
         return new Promise((res, rej) => {
             user.setUserMfaPreference(smsMfaSettings, totpMfaSettings, (err, result) => {
                 if (err) {
-                    // if totp not setup or verified and user want to set it, return error
-                    // otherwise igonre it
-                    if (err.message === TOTP_NOT_SETUP || err.message === TOTP_NOT_VERIFED) {
-                        if (mfaMethod === 'SMS') {
-                            that.enableSMS(user).then((data) => {
-                                logger.debug('Set user mfa success', data);
-                                res(data);
-                            }).catch(err => {
-                                logger.debug('Set user mfa preference error', err);
-                                rej(err);
-                            });
-                        } else if (mfaMethod === 'NOMFA') {
-                            // diable sms
-                            that.disableSMS(user).then((data) => {
-                                logger.debug('Set user mfa success', data);
-                                res(data);
-                            }).catch(err => {
-                                logger.debug('Set user mfa preference error', err);
-                                rej(err);
-                            });
-                        } else {
-                            logger.debug('Set user mfa preference error', err);
-                            rej(err);
-                        }
-                    } else {
-                        logger.debug('Set user mfa preference error', err);
-                        rej(err);
-                    }
+                    logger.debug('Set user mfa preference error', err);
+                    rej(err);
+                    return;
                 }
                 logger.debug('Set user mfa success', result);
                 res(result);
+                return;
             });
         });
     }
 
     /**
      * diable SMS
+     * @deprecated
      * @param {CognitoUser} user - the current user
      * @return - A promise resolves is success
      */
@@ -525,15 +589,18 @@ export default class AuthClass {
                 if (err) {
                     logger.debug('disable mfa failed', err);
                     rej(err);
+                    return;
                 }
                 logger.debug('disable mfa succeed', data);
                 res(data);
+                return;
             });
         });
     }
 
     /**
      * enable SMS
+     * @deprecated
      * @param {CognitoUser} user - the current user
      * @return - A promise resolves is success
      */
@@ -543,9 +610,11 @@ export default class AuthClass {
                 if (err) {
                     logger.debug('enable mfa failed', err);
                     rej(err);
+                    return;
                 }
                 logger.debug('enable mfa succeed', data);
                 res(data);
+                return;
             });
         });
     }
@@ -561,10 +630,12 @@ export default class AuthClass {
                 onFailure: (err) => {
                     logger.debug('associateSoftwareToken failed', err);
                     rej(err);
+                    return;
                 },
                 associateSecretCode: (secretCode) => {
                     logger.debug('associateSoftwareToken sucess', secretCode);
                     res(secretCode);
+                    return;
                 }
             });
         });
@@ -583,10 +654,12 @@ export default class AuthClass {
                 onFailure: (err) => {
                     logger.debug('verifyTotpToken failed', err);
                     rej(err);
+                    return;
                 },
                 onSuccess: (data) => {
                     logger.debug('verifyTotpToken success', data);
                     res(data);
+                    return;
                 }
             });
         });
@@ -709,7 +782,13 @@ export default class AuthClass {
                         }
                     }
                     user.updateAttributes(attributeList, (err,result) => {
-                        if (err) { reject(err); } else { resolve(result); }
+                        if (err) { 
+                            reject(err); 
+                            return;
+                        } else { 
+                            resolve(result); 
+                            return;
+                        }
                     });
                 });
             });
@@ -774,8 +853,44 @@ export default class AuthClass {
                     return;
                 }
 
+                // refresh the session if the session expired.
                 user.getSession(function(err, session) {
-                    if (err) { rej(err); } else { res(user); }
+                    if (err) {
+                        logger.debug('Failed to get the user session', err);
+                        rej(err); 
+                        return;
+                    }
+                });
+
+                // get user data from Cognito
+                user.getUserData((err, data) => {
+                    if (err) {
+                        logger.debug('getting user data failed', err);
+                        // Make sure the user is still valid
+                        if (err.message === 'User is disabled' || err.message === 'User does not exist.') {
+                            rej(err);
+                        } else {
+                            // the error may also be thrown when lack of permissions to get user info etc
+                            // in that case we just bypass the error
+                            res(user);
+                        }
+                        return;
+                    }
+                    const preferredMFA = data.PreferredMfaSetting || 'NOMFA';
+                    const attributeList = [];
+
+                    for (let i = 0; i < data.UserAttributes.length; i++) {
+                        const attribute = {
+                            Name: data.UserAttributes[i].Name,
+                            Value: data.UserAttributes[i].Value,
+                        };
+                        const userAttribute = new CognitoUserAttribute(attribute);
+                        attributeList.push(userAttribute);
+                    }
+
+                    const attributes = this.attributesToObject(attributeList);
+                    Object.assign(user, {attributes, preferredMFA});
+                    res(user);
                 });
             });
         });
@@ -804,17 +919,11 @@ export default class AuthClass {
             try {
                 user = await this.currentUserPoolUser();
             } catch (e) {
-                throw 'not authenticated';
+                logger.debug('The user is not authenticated by the error', e);
+                throw ('not authenticated');
             }
-            let attributes = {};
-            try {
-                attributes = this.attributesToObject(await this.userAttributes(user));
-            } catch (e) {
-                logger.debug('cannot get user attributes');
-            } finally {
-                this.user = Object.assign(user, { attributes });
-                return this.user;
-            }
+            this.user = user;
+            return this.user;
         }
     }
 
@@ -860,10 +969,12 @@ export default class AuthClass {
             user.getSession(function(err, session) {
                 if (err) { 
                     logger.debug('Failed to get the session from user', user);
-                    reject(err); 
+                    reject(err);
+                    return;
                 } else {
                     logger.debug('Succeed to get the user session', session);
                     resolve(session); 
+                    return;
                 }
             });
         });
@@ -915,8 +1026,14 @@ export default class AuthClass {
     public verifyUserAttribute(user, attr): Promise<any> {
         return new Promise((resolve, reject) => {
             user.getAttributeVerificationCode(attr, {
-                onSuccess(data) { resolve(data); },
-                onFailure(err) { reject(err); }
+                onSuccess(data) { 
+                    resolve(data); 
+                    return;
+                },
+                onFailure(err) {
+                    reject(err); 
+                    return;
+                }
             });
         });
     }
@@ -933,8 +1050,14 @@ export default class AuthClass {
 
         return new Promise((resolve, reject) => {
             user.verifyAttribute(attr, code, {
-                onSuccess(data) { resolve(data); },
-                onFailure(err) { reject(err); }
+                onSuccess(data) { 
+                    resolve(data); 
+                    return;
+                },
+                onFailure(err) { 
+                    reject(err); 
+                    return;
+                }
             });
         });
     }
@@ -1014,8 +1137,10 @@ export default class AuthClass {
                         if (err) {
                             logger.debug('change password failure', err);
                             reject(err);
+                            return;
                         } else {
                             resolve(data);
+                            return;
                         }
                     });
                 });
@@ -1034,13 +1159,18 @@ export default class AuthClass {
         const user = this.createCognitoUser(username);
         return new Promise((resolve, reject) => {
             user.forgotPassword({
-                onSuccess: () => { resolve(); },
+                onSuccess: () => { 
+                    resolve();
+                    return; 
+                },
                 onFailure: err => {
                     logger.debug('forgot password failure', err);
                     reject(err);
+                    return;
                 },
                 inputVerificationCode: data => {
                     resolve(data);
+                    return;
                 }
             });
         });
@@ -1066,8 +1196,14 @@ export default class AuthClass {
         const user = this.createCognitoUser(username);
         return new Promise((resolve, reject) => {
             user.confirmPassword(code, password, {
-                onSuccess: () => { resolve(); },
-                onFailure: err => { reject(err); }
+                onSuccess: () => { 
+                    resolve(); 
+                    return;
+                },
+                onFailure: err => { 
+                    reject(err); 
+                    return;
+                }
             });
         });
     }
@@ -1130,8 +1266,10 @@ export default class AuthClass {
                 dispatchAuthEvent('signIn', that.user);
                 logger.debug('federated sign in credentials', cred);
                 res(cred);
+                return;
             }).catch(e => {
                 rej(e);
+                return;
             });
         });    
     }