+++ packages/aws-amplify-angular/src/components/index.ts
@@ -1,5 +1,0 @@
-export * from './storage';
-export * from './authenticator';
-export * from './common';
-export * from './interactions';
-export { ComponentMount } from './component.mount'